Extract some functions to utils file
